requirements.txt
tensorflow-text==2.12.1
tensorflow-datasets==4.9.2
tensorflow==2.12.1
python-dotenv==1.0.0
streamlit==1.25.0
regex==2023.8.8